move conftest.py to root directory of tests
